refactor: fetch 로직이 추가됨에 따라 장소 리스트 상세 조회 훅 분리
