+++ hooks/place-list/useGetPlaceListDetail.ts
@@ -1,61 +1,23 @@
-import { getPlaceListDetail, getPlaceListTags } from '@/lib/actions/placeList';
-import { PageParam, Place } from '@/types/placeList';
-import { infiniteQueryOptions, queryOptions, useSuspenseInfiniteQuery, useSuspenseQuery } from '@tanstack/react-query';
+import { RpcError } from '@/types/errors';
+import { PlaceListDetail } from '@/types/placeList';
+import { queryOptions, useSuspenseQuery } from '@tanstack/react-query';
+
+const fetchPlaceListDetail = async (listId: string): Promise<PlaceListDetail> => {
+  const res = await fetch(`/api/view/place-list/${listId}/detail`);
+  if (res.status === 401) throw new RpcError('UNAUTHORIZED');
+  if (!res.ok) throw new RpcError('장소 리스트 정보를 가져오는 데 실패했습니다.');
+  return res.json();
+};
 
 // queryOptions 정의 - 서버/클라이언트 캐시 동기화를 위해
 // 리스트 개요 queryOptions
 export const placeListDetailQueryOptions = (listId: string) => {
   return queryOptions({
     queryKey: ['place-list', listId, 'detail'],
-    queryFn: async () => {
-      const result = await getPlaceListDetail(listId);
-      if ('error' in result) {
-        throw new Error(result.code ?? result.error);
-      }
-      return result;
-    },
-  });
-};
-
-const fetchPlaceListPlaces = async (listId: string, cursor: PageParam): Promise<Place[]> => {
-  const params = new URLSearchParams();
-  if (cursor) {
-    params.set('createdAt', cursor.createdAt);
-    params.set('id', cursor.id);
-  }
-  const res = await fetch(`/api/view/place-list/${listId}/places?${params.toString()}`);
-  if (!res.ok) throw new Error('저장된 장소를 가져오는 데 실패했습니다.');
-  return res.json();
-};
-
-// 리스트에 저장된 장소 queryOptions
-export const placeListPlacesQueryOptions = (listId: string) => {
-  return infiniteQueryOptions({
-    queryKey: ['place-list', listId, 'places'],
-    queryFn: ({ pageParam }) => fetchPlaceListPlaces(listId, pageParam),
-    initialPageParam: null as PageParam,
-    getNextPageParam: (lastPage) =>
-      lastPage.length < 20 ? undefined : { createdAt: lastPage.at(-1)!.createdAt, id: lastPage.at(-1)!.id },
-  });
-};
-
-// 리스트에 저장된 태그
-export const placeListTagsQueryOptions = (listId: string) => {
-  return queryOptions({
-    queryKey: ['place-list', listId, 'tags'],
-    queryFn: () => getPlaceListTags(listId),
+    queryFn: () => fetchPlaceListDetail(listId),
   });
 };
 
 export const useGetPlaceListDetail = (listId: string) => {
   return useSuspenseQuery(placeListDetailQueryOptions(listId));
-};
-
-export const useGetPlaceListPlaces = (listId: string) => {
-  const { data, ...rest } = useSuspenseInfiniteQuery(placeListPlacesQueryOptions(listId));
-  return { data: data.pages.flat(), ...rest };
-};
-
-export const useGetPlaceListTags = (listId: string) => {
-  return useSuspenseQuery(placeListTagsQueryOptions(listId));
 };
